Fix linecoverage job
sudo -E env do show that on Fedora. /home is not reset, this breakline
coverage in weird way (georep test grepping in the wrong directory,
for a start). Tracking that down, it come from a difference of
configuration between Fedora and Centos, see
https://bugzilla.redhat.com/show_bug.cgi?id=1728687

Until that's fixed in Fedora, we have to fix it ourself.