Insert the line at the start of the file
By default, ansible would insert it at the end. And so
that mean the configuration is applied only on files to rotate
declared after all the others, thus applied to nothing.
This is not clearly documented in the man page,and global options
are global just by convention, aka appearing before the files.

This is likely why the log were not compressed, and why we did see
download server having issue with them.