Fix kernel upgrade woes
For a reason that I do not know (but that I tried to understand),
the kernel decided that eth0 is now eth1. All attempts to rename that
were met with problem, until I renamed them to eth3/4 (since there is
no race condition).

In turn, this broke some assumption around keepalived, assuming the
same interface name was used on both side. Then I discovered that
vars set at the role level were not present in hostvars, but if
I copy them using set_fact, they do appear.

Ergo, this commit.