Separate cloud instance installation
Due to a bug in the role, it seems that we are adding
only 1 single VM for each type in jenkins.
I have not yet found what is the problem, but I suspect
that doing a separate play would work fine for that.

The bug can be seen because jenkins try to add the same VM twice, and
fail, and because fedora setup seems to use the centos value
for fact. I guess it might be a issue with multiplerole inclusion
somewhere down the line.