Fix server installation
Seems the use of set_fact in aws_builder_c7 mess with the aws_builder_f37
role, by overriding a variable. It was likely a ansible change, so this
part may need to be fixed in ansible, or in our role. In the mean time,
it should be enough to move that at the bottom of the playbook